requirements.txt
streamlit
opencv-python-headless
Pillow
tensorflow
fastai
torch
langchain
ctransformers
sentencepiece
scikit-learn
numpy
gdown